Created Feature to Remove Notes
- Once auto notes has been hit, the next time
it will remove the notes, this is so people
can turn on or turn off notes at will. I know this
introduces the bug that if we hit the auto notes
button when theres no notes but then we've deleted
the notes manually, we will have to double click.
I can fix this edge case later. For now I have to fix
the bug that once I undo right after a auto candidates move
i should have to undo the autoMode to previous setting.